Create xccdf package in openscap-plugin
Moved XCCDF related code from server.go to new files in xccdf package.
Also refactored the existing functions related to tailoring file.
