+++ cmd/openscap-plugin/server/server.go
@@ -5,7 +5,6 @@ package server
 
 import (
 	"bufio"
-	"encoding/xml"
 	"errors"
 	"fmt"
 	"os"
@@ -14,12 +13,12 @@ import (
 	"time"
 
 	"github.com/ComplianceAsCode/compliance-operator/pkg/utils"
-	"github.com/ComplianceAsCode/compliance-operator/pkg/xccdf"
 	"github.com/antchfx/xmlquery"
 	"github.com/oscal-compass/compliance-to-policy-go/v2/policy"
 
 	"github.com/complytime/complytime/cmd/openscap-plugin/config"
 	"github.com/complytime/complytime/cmd/openscap-plugin/scan"
+	"github.com/complytime/complytime/cmd/openscap-plugin/xccdf"
 )
 
 var _ policy.Provider = (*PluginServer)(nil)
@@ -36,7 +35,7 @@ func New(cfg *config.Config) PluginServer {
 
 func (s PluginServer) Generate(policy policy.Policy) error {
 	fmt.Println("Generating a tailoring file")
-	tailoringXML, err := policyToXML(policy, s.Config)
+	tailoringXML, err := xccdf.PolicyToXML(policy, s.Config)
 	if err != nil {
 		return err
 	}
@@ -73,7 +72,7 @@ func (s PluginServer) GetResults(_ policy.Policy) (policy.PVPResult, error) {
 		return policy.PVPResult{}, err
 	}
 
-	ruleTable := newRuleHashTable(xmlnode)
+	ruleTable := xccdf.NewRuleHashTable(xmlnode)
 	results := xmlnode.SelectElements("//rule-result")
 	for i := range results {
 		result := results[i]
@@ -139,101 +138,4 @@ func mapResultStatus(result *xmlquery.Node) (policy.Result, error) {
 	}
 
 	return policy.ResultInvalid, fmt.Errorf("couldn't match %s", resultEl.InnerText())
-}
-
-func policyToXML(tp policy.Policy, config *config.Config) (string, error) {
-	tailoring := xccdf.TailoringElement{
-		XMLNamespaceURI: xccdf.XCCDFURI,
-		ID:              getTailoringID(),
-		Version: xccdf.VersionElement{
-			Time:  time.Now().Format(time.RFC3339),
-			Value: "1",
-		},
-		Benchmark: xccdf.BenchmarkElement{
-			Href: config.Files.Datastream,
-		},
-		Profile: xccdf.ProfileElement{
-			ID: GetXCCDFProfileID(),
-			Title: &xccdf.TitleOrDescriptionElement{
-				Value: "example",
-			},
-			Selections: getSelections(tp),
-			Values:     getValuesFromVariables(tp),
-		},
-	}
-
-	output, err := xml.MarshalIndent(tailoring, "", "  ")
-	if err != nil {
-		return "", err
-	}
-	return xccdf.XMLHeader + "\n" + string(output), nil
-}
-
-func getSelections(tp policy.Policy) []xccdf.SelectElement {
-	var selections []xccdf.SelectElement
-	for _, rule := range tp {
-		selections = append(selections, xccdf.SelectElement{
-			IDRef:    rule.Rule.ID,
-			Selected: true,
-		})
-	}
-	// Disable all else?
-	return selections
-}
-
-func getValuesFromVariables(tp policy.Policy) []xccdf.SetValueElement {
-	var values []xccdf.SetValueElement
-	if len(tp) != 0 {
-		for _, rule := range tp {
-			if rule.Rule.Parameter == nil {
-				continue
-			}
-
-			values = append(values, xccdf.SetValueElement{
-				IDRef: rule.Rule.Parameter.ID,
-				Value: rule.Rule.Parameter.Value,
-			})
-		}
-		return values
-	}
-
-	return nil
-}
-
-// GetXCCDFProfileID gets a profile xccdf ID from the TailoredProfile object
-func GetXCCDFProfileID() string {
-	return fmt.Sprintf("xccdf_%s_profile_%s", xccdf.XCCDFNamespace, "my-tailoring-profile")
-}
-
-func getTailoringID() string {
-	return fmt.Sprintf("xccdf_%s_tailoring_%s", xccdf.XCCDFNamespace, "my-tailoring-profile")
-}
-
-// Getting rule information
-// Copied from https://github.com/ComplianceAsCode/compliance-operator/blob/fed54b4b761374578016d79d97bcb7636bf9d920/pkg/utils/parse_arf_result.go#L170
-
-func newRuleHashTable(dsDom *xmlquery.Node) NodeByIdHashTable {
-	return newHashTableFromRootAndQuery(dsDom, "//ds:component/xccdf-1.2:Benchmark", "//xccdf-1.2:Rule")
-}
-
-func newHashTableFromRootAndQuery(dsDom *xmlquery.Node, root, query string) NodeByIdHashTable {
-	benchmarkDom := dsDom.SelectElement(root)
-	rules := benchmarkDom.SelectElements(query)
-	return newByIdHashTable(rules)
-}
-
-type NodeByIdHashTable map[string]*xmlquery.Node
-
-//type nodeByIdHashVariablesTable map[string][]string
-
-func newByIdHashTable(nodes []*xmlquery.Node) NodeByIdHashTable {
-	table := make(NodeByIdHashTable)
-	for i := range nodes {
-		ruleDefinition := nodes[i]
-		ruleId := ruleDefinition.SelectAttr("id")
-
-		table[ruleId] = ruleDefinition
-	}
-
-	return table
 }
