+++ cmd/openscap-plugin/xccdf/datastream.go
@@ -1,0 +1,36 @@
+// SPDX-License-Identifier: Apache-2.0
+
+package xccdf
+
+import (
+	"github.com/antchfx/xmlquery"
+)
+
+// Getting rule information
+// Copied from https://github.com/ComplianceAsCode/compliance-operator/blob/fed54b4b761374578016d79d97bcb7636bf9d920/pkg/utils/parse_arf_result.go#L170
+
+func NewRuleHashTable(dsDom *xmlquery.Node) NodeByIdHashTable {
+	return newHashTableFromRootAndQuery(dsDom, "//ds:component/xccdf-1.2:Benchmark", "//xccdf-1.2:Rule")
+}
+
+func newHashTableFromRootAndQuery(dsDom *xmlquery.Node, root, query string) NodeByIdHashTable {
+	benchmarkDom := dsDom.SelectElement(root)
+	rules := benchmarkDom.SelectElements(query)
+	return newByIdHashTable(rules)
+}
+
+type NodeByIdHashTable map[string]*xmlquery.Node
+
+//type nodeByIdHashVariablesTable map[string][]string
+
+func newByIdHashTable(nodes []*xmlquery.Node) NodeByIdHashTable {
+	table := make(NodeByIdHashTable)
+	for i := range nodes {
+		ruleDefinition := nodes[i]
+		ruleId := ruleDefinition.SelectAttr("id")
+
+		table[ruleId] = ruleDefinition
+	}
+
+	return table
+}
